pr_agent

Module pr_agent

here we will have node sre2_agent centered this will hold the node logic using prompts specific and structured output schema specific to any mini-node we have broken down the agents task into smaller ones easier to digest and compatible with local llm limitations so ‘conditional edges’ will be in the bigger function that coordinated the nodes

Structs

PrAgentHandler

Functions

pr_agent_node_work_orchestration

run_pull

then it will be able to pull work

run_read_and_select

The pr agent will pass through different steps read sre report and select the sre agent concerned to pick the right git branch for next step, pull work, create a report on work done. thisis the read and select: after having receive instructions from message transmitted having the report which will tell which agent is concerned

run_report

finally it will be creating a report so that next agent get a nice overview of what has been done